trippai_ai/requirements.txt
pandas
numpy
requests
pytrends
scipy
python-dotenv
prophet
fastapi
uvicorn
pydantic
